+++ sdk/highlight-py/highlight_io/integrations/__init__.py
@@ -1,0 +1,9 @@
+import abc
+
+
+class Integration(abc.ABC):
+    def enable(self):
+        raise NotImplementedError
+
+    def disable(self):
+        raise NotImplementedError