+++ sdk/highlight-node/CHANGELOG.md
@@ -10,3 +10,10 @@
 
 -   Updated dependencies
     -   highlight.run@4.6.0
+
+## 2.0.0
+
+### Major Changes
+
+-   require project id for H.init
+-   support for errors without associated sessions/requests